Removed adaptive image ajax script
The script is well written and works perfectly in localhost production
environment, but causes as 406 error on click in production server.
Opening the URL generated by GET, leads to an error page, which
describes mod_security as the culprit. Restricting mod_security is
likely to decrease security and require involvement from host.
Understanding why mod_security is causing this error might lead to a
solution. Removed for now.