Improved SASS syntax of home page styles
Grid is only applied to <main> and <aside> if body:not(.home) and
sections are only styled if body.home. The conditional styles for .home
are now embedded within the elements which they affect using & at the
end of the conditional SASS selector to prefix that conditional selector
to the CSS selector.